Generates meta domains instead of meta actions
