+++ MetaActions.Learn/MetaActions.Learn.csproj
@@ -9,6 +9,7 @@
 
   <ItemGroup>
     <PackageReference Include="CommandLineParser" Version="2.9.1" />
+    <PackageReference Include="PDDLSharp" Version="1.1.7" />
   </ItemGroup>
 
   <ItemGroup>
